requirements.txt
fastapi
uvicorn
python-dotenv
groq
chromadb
langchain
langchain-core
langchain-groq
langchain-community
langgraph
pypdf
httpx
requests
python-multipart
scikit-learn
twilio
webrtcvad
websockets
pydantic
jinja2
